지훈님 오류 수정 수정

+++ onezo3/src/main/java/com/onezogreen/onezo3/order/OrderVo.java
@@ -21,5 +21,5 @@ public class OrderVo {
     private Timestamp orderDate;
     private TakeInOut takeInOut;
     private List<OrderItemVo> orderItems;
-    private BigDecimal totalPrice;
+    private int totalPrice;
 }
